Post testing bugfixes. Fix speed calculation. Adjust Thresholds.
